Add files via upload
This project involves the design and implementation of a library database using SQL statements in Microsoft SQL Server Management Studio. The goal of the project was to create an efficient and easy to use database management system that will be used to manage the library items which includes a collection of books, journals, DVD and other media, as well as ensure effective management of member details and lending operations.