Create README.md
I really enjoyed working on this project as part of my course work as it demonstrated my proficiency in the use of Microsoft SQL Server Management Studio and SQL for database development. 
The file contains the script to create the different tables and also scripts create stored procedures, user-defined functions, views and triggers used to simplify library operations.